Minor cleanup, renaming
