+++ packages/cli/src/commands/destroy/component/component.js
@@ -1,9 +1,9 @@
-import { createYargsForComponentDestroy } from '../../generate/helpers'
-import { files } from '../../generate/component/component'
+import { files as componentFiles } from '../../generate/component/component'
+import { createYargsForComponentDestroy } from '../helpers'
 
 export const desc = 'Destroy a component.'
 
 export const { command, builder, handler } = createYargsForComponentDestroy({
   componentName: 'component',
-  filesFn: files,
+  filesFn: componentFiles,
 })
