+++ packages/cli/src/commands/destroy/page/page.js
@@ -5,7 +5,7 @@ import { deleteFilesTask, removeRoutesFromRouterTask } from 'src/lib'
 import c from 'src/lib/colors'
 
 import { pathName } from '../../generate/helpers'
-import { files } from '../../generate/page/page'
+import { files as pageFiles } from '../../generate/page/page'
 
 export const command = 'page <name> [path]'
 export const desc = 'Destroy a page component.'
@@ -16,7 +16,7 @@ export const handler = async ({ name, path }) => {
       {
         title: 'Destroying page files...',
         task: async () => {
-          const f = files({ name, path: pathName(path, name) })
+          const f = pageFiles({ name, path: pathName(path, name) })
           return deleteFilesTask(f)
         },
       },
